requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
Pillow>=10.0.0
python-dotenv>=1.0.0
drf-nested-routers
# Development dependencies
ipdb>=0.13.13
django-extensions>=4.2.0
django-debug-toolbar>=4.2.0
pylint-django>=2.5.3
black>=23.11.0
isort>=5.12.0
